// Projects/Tamin/Eservices-UI/src/app/models/registration/personal.model.ts
export class PersonalModel {
  data: PersonalModel;
  parentId: PersonalModel;
  refrenceCode: string;
  id: number;
  firstName: string;
  lastName: string;
  fatherName: string;
  countryId: string;
  idCardSerial1: string;
  idCardSerial2: string;
  ssn: string;
  city: any;
  cityTitle: string;
  BirtDateCity: string;
  dateOfBirth: string;
  CreatedCity: string;
  birthDate: string;
  gender: any;
  nation: string;
  idCardNumber: string;
  country: any;
  cityOfBirthId: any;
  cityOfIssueId: any;
  cityOfBirthName: string;
  cityOfIssueName: string;
  nationality: string;
  nationalId: string;
  dependentType: any;
  request: any;
  bailType: any;
  isNew:string;
}